backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pillow==10.4.0
python-multipart==0.0.12
aiofiles==24.1.0
python-dotenv==1.0.1